requirements.txt
# Core framework
Django==6.0.1

# Django REST Framework
djangorestframework==3.16.1

# JWT authentication
djangorestframework-simplejwt==5.3.1

# Swagger/OpenAPI docs
drf-yasg==1.21.14

# CORS headers (frontend integration)
django-cors-headers==4.3.1

# Optional utilities
django-extensions==3.2.3

# Supporting libraries
PyYAML==6.0.3
uritemplate==4.2.0
inflection==0.5.1
packaging==26.0
pytz==2025.2
sqlparse==0.5.5
asgiref==3.11.0
